refactor(core): extract llmChat backlink excerpts into a service module
